dac_platform/DAC/requirements.txt
aioredis==1.2.0
APScheduler==3.6.0
asgiref==2.3.2
async-timeout==2.0.1
attrs==19.1.0
autobahn==19.3.3
Automat==0.7.0
backports.csv==1.0.7
certifi==2019.3.9
channels==2.0.0
channels-redis==2.3.2
chardet==3.0.4
constantly==15.1.0
daphne==2.2.5
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.1.7
django-apscheduler==0.3.0
django-cors-headers==2.5.0
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.2.0
django-reversion==3.0.3
djangorestframework==3.9.2
et-xmlfile==1.0.1
future==0.17.1
futures==3.1.1
hiredis==1.0.0
httplib2==0.9.2
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
jdcal==1.4.1
msgpack==0.5.6
msgpack-python==0.5.6
odfpy==1.4.0
openpyxl==2.6.2
paho-mqtt==1.4.0
PyHamcrest==1.9.0
PyMySQL==0.9.3
pytz==2018.9
PyYAML==5.1
redis==2.10.6
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
tablib==0.13.0
Twisted==18.9.0
txaio==18.8.1
tzlocal==1.5.1
urllib3==1.24.1
xadmin2==2.0.1
xlrd==1.2.0
xlwt==1.3.0
zope.interface==4.6.0
